Revert "refactor(workers): split provider reconciliation"
This reverts commit 0a1f97399b23eca81e99af45dd9ea616e266de54.
